feat: created router and users
