+++ .eslintrc.json
@@ -20,6 +20,8 @@
     "@typescript-eslint/explicit-module-boundary-types": "off",
     "@typescript-eslint/no-explicit-any": "off",
     "@typescript-eslint/no-var-requires": "off",
-    "no-console": "off"
+    "no-console": "off",
+    "import/prefer-default-export": "off",
+    "import/extensions": "off"
   }
 }
